(#564) move zocaloresults to testing utils
